+++ src/main/java/subway/application/LineService.java
@@ -2,6 +2,10 @@ package subway.application;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import subway.application.request.LineRequest;
+import subway.application.request.SectionRequest;
+import subway.application.request.StationRequest;
+import subway.application.response.LineResponse;
 import subway.domain.line.Line;
 import subway.domain.line.LineName;
 import subway.domain.section.Distance;
@@ -9,10 +13,6 @@ import subway.domain.section.Section;
 import subway.domain.section.Sections;
 import subway.domain.station.Station;
 import subway.persistence.repository.LineRepository;
-import subway.ui.request.LineRequest;
-import subway.ui.request.SectionRequest;
-import subway.ui.request.StationRequest;
-import subway.ui.response.LineResponse;
 
 import java.util.List;
 import java.util.stream.Collectors;
