+++ src/main/java/subway/application/PathService.java
@@ -1,13 +1,13 @@
 package subway.application;
 
 import org.springframework.stereotype.Service;
+import subway.application.request.PathRequest;
+import subway.application.response.PathResponse;
 import subway.domain.path.Fare;
 import subway.domain.path.Path;
 import subway.domain.station.Station;
 import subway.domain.station.StationName;
 import subway.persistence.repository.PathRepository;
-import subway.ui.request.PathRequest;
-import subway.ui.response.PathResponse;
 
 import java.util.List;
