+++ src/main/java/subway/persistence/repository/StationRepository.java
@@ -1,10 +1,10 @@
 package subway.persistence.repository;
 
 import org.springframework.stereotype.Repository;
+import subway.application.request.StationRequest;
 import subway.domain.station.Station;
 import subway.persistence.dao.StationDao;
 import subway.persistence.entity.StationEntity;
-import subway.ui.request.StationRequest;
 
 import java.util.List;
 import java.util.stream.Collectors;
